Drawer Navigation, Navigation Stack Refactor, and Nav Header Branding (#1137)
* Drawer Navigation, Navigation Stack Refactor, and Nav Header Branding

Addressed PR feedback

* Restructure stack to remove RootStack

Fix bad merge
