+++ App.tsx
@@ -1,9 +1,7 @@
 import React, {useCallback, useEffect, useMemo, useState} from 'react';
 
-import Ionicons from '@expo/vector-icons/Ionicons';
 import {SelectProvider} from '@mobile-reality/react-native-select-pro';
-import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
-import {getStateFromPath, NavigationContainer, PathConfigMap, RouteProp, useNavigationContainerRef} from '@react-navigation/native';
+import {getStateFromPath, NavigationContainer, PathConfigMap, useNavigationContainerRef} from '@react-navigation/native';
 import * as SplashScreen from 'expo-splash-screen';
 import {ActivityIndicator, AppState, AppStateStatus, Image, Platform, StatusBar, StyleSheet, View} from 'react-native';
 import {SafeAreaProvider} from 'react-native-safe-area-context';
@@ -14,8 +12,6 @@ import * as BackgroundFetch from 'expo-background-task';
 import Constants from 'expo-constants';
 import * as TaskManager from 'expo-task-manager';
 
-import {merge} from 'lodash';
-
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import {createAsyncStoragePersister} from '@tanstack/query-async-storage-persister';
 import {focusManager, QueryCache, QueryClient, useQueryClient} from '@tanstack/react-query';
@@ -23,10 +19,6 @@ import {PersistQueryClientProvider} from '@tanstack/react-query-persist-client';
 
 import {ClientContext, ClientProps, productionHosts, stagingHosts} from 'clientContext';
 import {ActionToast, ErrorToast, InfoToast, SuccessToast, WarningToast} from 'components/content/Toast';
-import {HomeTabScreen} from 'components/screens/HomeScreen';
-import {MenuStackScreen} from 'components/screens/MenuScreen';
-import {ObservationsTabScreen} from 'components/screens/ObservationsScreen';
-import {WeatherScreen} from 'components/screens/WeatherScreen';
 import {HTMLRendererConfig} from 'components/text/HTML';
 import ImageCache, {queryKeyPrefix} from 'hooks/useCachedImageURI';
 import {useOnlineManager} from 'hooks/useOnlineManager';
@@ -35,8 +27,6 @@ import {logFilePath, logger} from 'logger';
 import {LoggerContext, LoggerProps} from 'loggerContext';
 import {prefetchAllActiveForecasts} from 'network/prefetchAllActiveForecasts';
 import Toast, {ToastConfigParams} from 'react-native-toast-message';
-import {TabNavigatorParamList} from 'routes';
-import {colorLookup} from 'theme';
 import {AvalancheCenterWebsites} from 'types/nationalAvalancheCenter';
 
 import 'date-time-format-timezone';
@@ -46,7 +36,7 @@ import {QUERY_CACHE_ASYNC_STORAGE_KEY} from 'data/asyncStorageKeys';
 import * as FileSystem from 'expo-file-system';
 import {PreferencesProvider, usePreferences} from 'Preferences';
 import {NotFoundError} from 'types/requests';
-import {formatRequestedTime, RequestedTime} from 'utils/date';
+import {RequestedTime} from 'utils/date';
 
 import Mapbox from '@rnmapbox/maps';
 import {Integration} from '@sentry/types';
@@ -55,6 +45,7 @@ import * as messages from 'compiled-lang/en.json';
 import {Button} from 'components/content/Button';
 import {Center, VStack} from 'components/core';
 import {KillSwitchMonitor} from 'components/KillSwitchMonitor';
+import {DrawerNavigator} from 'components/screens/navigation/Drawer';
 import {Body, BodyBlack, Title3Black} from 'components/text';
 import * as Linking from 'expo-linking';
 import * as Updates from 'expo-updates';
@@ -237,8 +228,6 @@ const asyncStoragePersister = createAsyncStoragePersister({
   key: QUERY_CACHE_ASYNC_STORAGE_KEY,
 });
 
-const TabNavigator = createBottomTabNavigator<TabNavigatorParamList>();
-
 // We add the listener at startup and never plan to stop listening, so there's
 // no need to unsubscribe here.
 AppState.addEventListener('change', (status: AppStateStatus) => {
@@ -367,27 +356,6 @@ const BaseApp: React.FunctionComponent<{
       });
   }, [setUpdateStatus, logger]);
 
-  const tabNavigatorScreenOptions = useCallback(
-    ({route: {name}}: {route: RouteProp<TabNavigatorParamList, keyof TabNavigatorParamList>}) => ({
-      headerShown: false,
-      tabBarIcon: ({color, size}: {focused: boolean; color: string; size: number}) => {
-        if (name === 'Home') {
-          return <Ionicons name="map-outline" size={size} color={color} />;
-        } else if (name === 'Observations') {
-          return <Ionicons name="reader-outline" size={size} color={color} />;
-        } else if (name === 'Weather Data') {
-          return <Ionicons name="stats-chart-outline" size={size} color={color} />;
-        } else if (name === 'Menu') {
-          return <Ionicons name="ellipsis-horizontal" size={size} color={color} />;
-        }
-      },
-      // these two properties should really take ColorValue but oh well
-      tabBarActiveTintColor: colorLookup('primary') as string,
-      tabBarInactiveTintColor: colorLookup('text.secondary') as string,
-    }),
-    [],
-  );
-
   const [startupPaused, {off: unpauseStartup}] = useToggle(process.env.EXPO_PUBLIC_PAUSE_ON_STARTUP === 'true');
 
   if (updateStatus !== 'ready' || preferences.mixpanelUserId == '') {
@@ -514,52 +482,9 @@ const BaseApp: React.FunctionComponent<{
               <FeatureFlagsProvider>
                 <KillSwitchMonitor>
                   <SelectProvider>
-                    <StatusBar barStyle="dark-content" backgroundColor="white" />
+                    <StatusBar barStyle={'dark-content'} animated={false} backgroundColor={'white'} />
                     <View style={{flex: 1}}>
-                      <TabNavigator.Navigator initialRouteName="Home" screenOptions={tabNavigatorScreenOptions}>
-                        <TabNavigator.Screen name="Home" initialParams={{requestedTime: formatRequestedTime(requestedTime)}} options={{title: 'Zones'}}>
-                          {state =>
-                            HomeTabScreen(
-                              merge(state, {
-                                route: {
-                                  params: {
-                                    requestedTime: formatRequestedTime(requestedTime),
-                                  },
-                                },
-                              }),
-                            )
-                          }
-                        </TabNavigator.Screen>
-                        <TabNavigator.Screen name="Observations" initialParams={{requestedTime: formatRequestedTime(requestedTime)}}>
-                          {state =>
-                            ObservationsTabScreen(
-                              merge(state, {
-                                route: {
-                                  params: {
-                                    requestedTime: formatRequestedTime(requestedTime),
-                                  },
-                                },
-                              }),
-                            )
-                          }
-                        </TabNavigator.Screen>
-                        <TabNavigator.Screen name="Weather Data" initialParams={{requestedTime: formatRequestedTime(requestedTime)}}>
-                          {state =>
-                            WeatherScreen(
-                              merge(state, {
-                                route: {
-                                  params: {
-                                    requestedTime: formatRequestedTime(requestedTime),
-                                  },
-                                },
-                              }),
-                            )
-                          }
-                        </TabNavigator.Screen>
-                        <TabNavigator.Screen name="Menu" initialParams={{requestedTime: formatRequestedTime(requestedTime)}} options={{title: 'More'}}>
-                          {state => MenuStackScreen(state, queryCache, staging, setStaging)}
-                        </TabNavigator.Screen>
-                      </TabNavigator.Navigator>
+                      <DrawerNavigator requestedTime={requestedTime} centerId={avalancheCenterId} staging={staging} setStaging={setStaging} />
                     </View>
                   </SelectProvider>
                 </KillSwitchMonitor>
