+++ components/screens/main/AboutScreen.tsx
@@ -13,14 +13,15 @@ import Ionicons from '@expo/vector-icons/Ionicons';
 import {useFocusEffect} from '@react-navigation/native';
 import {ActionList} from 'components/content/ActionList';
 import {Center, HStack, View, VStack} from 'components/core';
-import {getVersionInfoFull} from 'components/screens/menu/Version';
+import {getVersionInfoFull} from 'components/screens/main/Version';
 import {Body, BodyBlack, BodyXSm, Title3Black} from 'components/text';
 import {getUpdateGroupId} from 'hooks/useEASUpdateStatus';
 import {usePostHog} from 'posthog-react-native';
 import {usePreferences} from 'Preferences';
-import {MenuStackParamList} from 'routes';
+import {useSafeAreaInsets} from 'react-native-safe-area-context';
+import {MainStackParamList} from 'routes';
 
-export const AboutScreen = (_: NativeStackScreenProps<MenuStackParamList, 'about'>) => {
+export const AboutScreen = (_: NativeStackScreenProps<MainStackParamList, 'about'>) => {
   const {
     preferences: {mixpanelUserId},
   } = usePreferences();
@@ -34,6 +35,8 @@ export const AboutScreen = (_: NativeStackScreenProps<MenuStackParamList, 'about
 
   const postHog = usePostHog();
 
+  const insets = useSafeAreaInsets();
+
   const recordAnalytics = useCallback(() => {
     postHog?.screen('about');
   }, [postHog]);
@@ -41,7 +44,7 @@ export const AboutScreen = (_: NativeStackScreenProps<MenuStackParamList, 'about
 
   return (
     <View style={StyleSheet.absoluteFillObject}>
-      <VStack backgroundColor="white" width="100%" height="100%" pt={16} justifyContent="space-between">
+      <VStack backgroundColor="white" width="100%" height="100%" pt={16} justifyContent="space-between" paddingBottom={insets.bottom}>
         <VStack space={16}>
           <Center>
             <Title3Black>About Avy</Title3Black>
