+++ package.json
@@ -50,6 +50,7 @@
     "@react-native/babel-preset": "*",
     "@react-native/normalize-color": "^2.0.0",
     "@react-navigation/bottom-tabs": "^7.10.1",
+    "@react-navigation/drawer": "^7.9.2",
     "@react-navigation/material-top-tabs": "^7.4.13",
     "@react-navigation/native": "^7.1.28",
     "@react-navigation/native-stack": "^7.11.0",
